requirements copy.txt
asyncpg==0.30.0
python-dotenv==1.0.1
langchain==0.3.26
langchain-google-genai==2.1.5
langchain-openai==0.3.27
langchain-mistralai==0.2.10
langchain-cohere==0.4.4
cohere==5.15.0
openai==1.93.0
mistralai==1.8.2
fastapi==0.115.14
uvicorn==0.34.3
